+++ src/controllers/user.controllers.js
@@ -1,7 +1,7 @@
-const db = require("../models");
-const sequelizeConfig = require('../config/sequelize.config')
-const User = db.user;
-const Op = db.Sequelize.Op;
+const db = require("../../db/models");
+const sequelizeConfig = require('../../config/sequelize.config')
+const User = db.User;
+const { v4 } = require('uuid')
 
 // Create and Save a new User
 // Method:POST, Endpoint:/user
@@ -14,28 +14,27 @@ exports.create = (req, res) => {
         return;
     }
 
+    if (!req.isAdmin) {
+        res.status(403).send({
+            message: "Forbidden",
+        });
+        return;
+    }
+    const { firstname, lastname, password, password2, email, roles } = req.body
+
     const passwordService = require('../services/password.services')
-    if (!passwordService.match(req.body.password, req.body.password2)) {
+    if (!passwordService.match(password, password2)) {
         res.status(400).send({
             message: "Password do not match",
         });
         return;
     }
-
-    const encryptedPassword = passwordService.encrypt(req.body.password)
-
-    // Create a User
-    const user = {
-        firstname: req.body.firstname,
-        lastname: req.body.lastname,
-        isActive: 0,
-        isChoiceAllowed: 0,
-        email: req.body.email,
-        password: encryptedPassword,
-        roles: req.body.roles,
-    };
+    const encryptedPassword = passwordService.encrypt(password)
 
     // Save User in the database
+    const uuid = v4()
+    const user = { uuid: uuid, firstname, lastname, isActive: 0, isChoiceAllowed: 0, email, password: encryptedPassword, roles }
+
     User.create(user)
         .then(data => {
             res.send(data);
@@ -51,8 +50,14 @@ exports.create = (req, res) => {
 // Retrieve all users
 // Method:GET, Endpoint:/users
 exports.findAll = (req, res) => {
+    if (!req.isAdmin) {
+        res.status(403).send({
+            message: "Forbidden",
+        });
+        return;
+    }
     User.findAll({
-        attributes: sequelizeConfig.models.user.attributes
+        attributes: sequelizeConfig.attributes.user
     }).then(data => {
         res.send(data);
     }).catch(err => {
@@ -63,12 +68,18 @@ exports.findAll = (req, res) => {
     });
 };
 
-// Retrieve one user from id
-// Method:GET, Endpoint:/user/:id
+// Retrieve one user from uuid
+// Method:GET, Endpoint:/user/:uuid
 exports.findOneById = (req, res) => {
+    if (!req.isAdmin) {
+        res.status(403).send({
+            message: "Forbidden",
+        });
+        return;
+    }
     User.findOne({
-        where: { id: req.params.id },
-        attributes: sequelizeConfig.models.user.attributes
+        where: { uuid: req.params.id },
+        attributes: sequelizeConfig.attributes.user
     }).then(data => {
         res.send(data);
     }).catch(err => {
@@ -80,19 +91,28 @@ exports.findOneById = (req, res) => {
 };
 
 // Update a User by the id in the request
-// Method:PUT, Endpoint:/user/:id
-exports.update = (req, res) => {
-    const user = {
-        firstname: req.body.firstname,
-        lastname: req.body.lastname,
-        isActive: req.body.isActive,
-        email: req.body.email,
-        roles: req.body.roles
-    };
+// Method:PUT, Endpoint:/user/:uuid
+exports.update = async (req, res) => {
+    const currentUser = await  User.findOne({
+        where: { id: req.user.id }
+    })
+
+    if (!req.isAdmin && currentUser.uuid !== req.params.id) {
+        res.status(403).send({
+            message: "Forbidden",
+        });
+        return;
+    }
+
+    const { firstname, lastname, isActive, email, roles } = req.body
+
+    const user = req.isAdmin ?
+        { firstname, lastname, isActive, email, roles } :
+        { firstname, lastname, email }
 
     User.update(
         user,
-        {returning: true, where: { id: req.params.id }}
+        {returning: true, where: { uuid: req.params.id }}
     ).then(data => {
         res.send(data);
     }).catch(err => {
@@ -104,10 +124,17 @@ exports.update = (req, res) => {
 };
 
 // Delete a User with the specified id in the request
-// Method:DELETE, Endpoint:/user/:id
+// Method:DELETE, Endpoint:/user/:uuid
 exports.deleteOne = (req, res) => {
+    if (!req.isAdmin) {
+        res.status(403).send({
+            message: "Forbidden",
+        });
+        return;
+    }
+
     User.destroy({
-        where: {id: req.params.id}
+        where: {uuid: req.params.id}
     }).then(data => {
         res.send(data);
     }).catch(err => {
